Beginning the cleaning process

+++ MaxText/layers/simple_decoder_layer.py
@@ -15,4 +15,4 @@ class SimpleDecoderLayer(nn.Module):
     self.weight_mat = self.param('weights', nn.initializers.ones, (self.config.emb_dim, self.config.emb_dim))
 
   def __call__(self, inputs: jnp.ndarray, positions, segmentation, deterministic, model_mode) -> jnp.ndarray:
-    return inputs @ self.weight_mat
+    return inputs @ self.weight_mat.astype(inputs.dtype)
